# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/group/create": {
            "post": {
                "tags": [
                    "群"
                ],
                "summary": "创建群组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "初始群员id,应包括群主",
                        "name": "members",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/group/delete": {
            "post": {
                "tags": [
                    "群"
                ],
                "summary": "删除群聊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "群id",
                        "name": "groupId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/group/enterAgree": {
            "post": {
                "tags": [
                    "群"
                ],
                "summary": "同意群申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "群id",
                        "name": "groupId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "被同意id",
                        "name": "agreedId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/group/enterReq": {
            "post": {
                "tags": [
                    "群"
                ],
                "summary": "发送加群申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "群id",
                        "name": "groupId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/group/getGroup": {
            "get": {
                "tags": [
                    "群"
                ],
                "summary": "获取群信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "群id",
                        "name": "groupId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/group/getMembers": {
            "get": {
                "tags": [
                    "群"
                ],
                "summary": "获取群员信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "群id",
                        "name": "groupId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/group/inviteMember": {
            "post": {
                "tags": [
                    "群"
                ],
                "summary": "邀请入群",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "群id",
                        "name": "groupId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "被邀请用户id",
                        "name": "invitedMembers",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/group/removeMember": {
            "post": {
                "tags": [
                    "群"
                ],
                "summary": "删除群员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "群id",
                        "name": "groupId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "被删除用户id",
                        "name": "deletedId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/message/getPush": {
            "get": {
                "tags": [
                    "消息操作"
                ],
                "summary": "创建通知通道",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.RespJson": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                },
                "data": {},
                "msg": {}
            }
        }
    }
}